added collections to PersonalSearchHit

+++ kifi-backend/modules/common/app/com/keepit/serializer/ArticleSearchResultSerializer.scala
@@ -52,7 +52,7 @@ class ArticleSearchResultSerializer extends Format[ArticleSearchResult] {
       (json \ "score").as[Float],
       (json \ "isMyBookmark").as[Boolean],
       (json \ "isPrivate").as[Boolean],
-      Seq((json \ "users").asInstanceOf[JsArray].value map {j => Id[User](j.as[Int])}: _*),
+      Seq((json \ "users").asInstanceOf[JsArray].value map {j => Id[User](j.as[Long])}: _*),
       (json \ "bookmarkCount").as[Int]
     )
   }
